Added LvlUp System + Fixed some bugs
-Run now has a chance to fail

+++ Rogue-Project/Assets/Scripts/Fighting/Classes/EnemyBaseClass.cs
@@ -37,7 +37,7 @@ public class EnemyBaseClass
     public float currentDefense;
 
     //EXP
-    public float expGiven;
+    public int expGiven;
 
     //Sprites
     public string spritePath;
